Poisson equation
The sample code in Python solves a Poisson equation (heat equation) in 1D using the finite element method.
where u(x,t) is the unknown function to be solved for, x is a coordinate in space, and t is time. The coefficient α is the diffusion coefficient and determines how fast u change in time. A quick short form for the diffusion equation is ut=αuxx ...